added default values to the equipment form

+++ src/Functions/Testing/demographicsValidation.ts
@@ -8,7 +8,7 @@ const validateAge = (age:number) => {
 const validateHeight = (height:number) => {
     //assumes inches
     if (height < 36 || height >= 96) throw new RangeError('Height is Outside of Bounds')
-    if (typeof height !== 'number') return parseInt(height)
+    if (typeof height !== 'number') return parseFloat(height)
     return height
 }
 
@@ -19,13 +19,13 @@ const validateSex = (sex: Sex) => {
 
 const validateCurrentWeight = (weight: number) => {
     if (weight < 80) throw new RangeError('Weight Is Outside Of Acceptable Bounds')
-    if (typeof weight !== 'number') return parseInt(weight)
+    if (typeof weight !== 'number') return parseFloat(weight)
     return weight
 }
 
 const validateGoalWeight = (weight:number) => {
     if (weight < 80 || weight > 400) throw new RangeError('Weight Is Outside Of Acceptable Bounds')
-    if (typeof weight !== 'number') return parseInt(weight)
+    if (typeof weight !== 'number') return parseFloat(weight)
     return weight
 }
